Make batch tracking information generic
Implements a generic tracking identifier for batches as they flow
through the submitter. This decouples the submitter from the
implementation of a batch identifier.

This allows the tracking ID to 1) change without impacting the
submitter's implementation, and 2) contain any sort of identifier or
metadata we would like.

The current implementation of the `BatchTrackingId` is a struct
containing the `batch_header` and the `service_id`.

This commit also updates the tests to conform to the new generic
pattern.

Resolves: #1385

Signed-off-by: Chris Eckhardt <eckhardt@bitwise.io>