Maybe a dead end
This is an attempt to use different structs for batches with service_id
vs no service_id with the submitter's existing components. This gets
very complicated.

We should probably pull many of the implementations here back into the
queuer, so that only a "sealed" batch envelope is passing through the
submitter. That will avoid a couple generics and box uses that make this
attempt messy.